+++ picodi/__init__.py
@@ -2,7 +2,7 @@ from picodi._picodi import (
     Provide,
     init_resources,
     inject,
-    make_dependency,
+    registry,
     resource,
     shutdown_resources,
 )
@@ -10,8 +10,8 @@ from picodi._picodi import (
 __all__ = [
     "init_resources",
     "inject",
-    "make_dependency",
     "Provide",
+    "registry",
     "resource",
     "shutdown_resources",
 ]