+++ tests/test_standard_dep_with_return.py
@@ -14,50 +14,72 @@ async def get_random_int_async():
     return random.randint(1, 100_000)
 
 
-@inject
-def get_redis_string(port: int = Provide(get_random_int)):
-    return f"http://redis:{port}"
+@pytest.fixture()
+def get_redis_string_dep():
+    @inject
+    def get_redis_string(port: int = Provide(get_random_int)):
+        return f"http://redis:{port}"
+
+    return get_redis_string
 
 
-@inject
-async def get_redis_string_async(port: int = Provide(get_random_int_async)):
-    return f"http://redis:{port}"
+@pytest.fixture()
+def get_redis_string_async_dep():
+    @inject
+    async def get_redis_string_async(port: int = Provide(get_random_int_async)):
+        return f"http://redis:{port}"
+
+    return get_redis_string_async
 
 
-@inject
-async def get_redis_string_async_with_sync_dep(port: int = Provide(get_random_int)):
-    return f"http://redis:{port}"
+@pytest.fixture()
+def get_sync_dependency_in_async_context_dep():
+    @inject
+    async def get_sync_dependency_in_async_context(port: int = Provide(get_random_int)):
+        return f"http://redis:{port}"
+
+    return get_sync_dependency_in_async_context
 
 
-def test_resolve_sync_dependency():
-    result = get_redis_string()
+def test_resolve_sync_dependency(get_redis_string_dep):
+    result = get_redis_string_dep()
 
     _check_redis_string(result)
 
 
-async def test_resolve_async_dependency():
-    result = await get_redis_string_async()
+async def test_resolve_async_dependency(get_redis_string_async_dep):
+    result = await get_redis_string_async_dep()
 
     _check_redis_string(result)
 
 
-async def test_resolve_sync_dependency_in_async_function():
-    result = await get_redis_string_async_with_sync_dep()
+async def test_resolve_sync_dependency_in_async_function(
+    get_sync_dependency_in_async_context_dep,
+):
+    result = await get_sync_dependency_in_async_context_dep()
 
     _check_redis_string(result)
 
 
-def test_resolve_dependency_multiple_times_return_different_results():
-    results = [get_redis_string() for _ in range(30)]
+def test_resolve_dependency_multiple_times_return_different_results(
+    get_redis_string_dep,
+):
+    results = [get_redis_string_dep() for _ in range(30)]
 
     assert len(set(results)) > 1
     _check_redis_string(results[0])
 
 
 @pytest.mark.parametrize(
-    "func", [get_redis_string_async, get_redis_string_async_with_sync_dep]
+    "func_name",
+    ["get_redis_string_async_dep", "get_sync_dependency_in_async_context_dep"],
 )
-async def test_resolve_async_dependency_multiple_times_return_different_results(func):
+async def test_resolve_async_dependency_multiple_times_return_different_results(
+    func_name,
+    get_redis_string_async_dep,  # noqa: U100
+    get_sync_dependency_in_async_context_dep,  # noqa: U100
+):
+    func = locals()[func_name]
     results = [await func() for _ in range(30)]
 
     assert len(set(results)) > 1
@@ -75,14 +97,14 @@ def test_resolve_async_dependency_from_sync_function_return_coroutine():
     assert result.__name__ == "get_random_int_async"
 
 
-def test_can_pass_dependency():
-    result = get_redis_string(port=100_000_000)
+def test_can_pass_dependency(get_redis_string_dep):
+    result = get_redis_string_dep(port=100_000_000)
 
     assert result == "http://redis:100000000"
 
 
-async def test_can_pass_dependency_async():
-    result = await get_redis_string_async(port=100_000_000)
+async def test_can_pass_dependency_async(get_redis_string_async_dep):
+    result = await get_redis_string_async_dep(port=100_000_000)
 
     assert result == "http://redis:100000000"